requirements.txt
selenium==4.20.0
python-dotenv==1.0.1
img2pdf==0.5.1